Add files via upload
This is a Continuous Performance Test program written in Turbo Pascal and is working under DOS.
You should first install DOSBox to your computer and then drag CPT.EXE to DOSBox in order to run. 
File names must be a maximum of 8 characters long without space. If you give the same file name for the new test, it will overwrite the old files without warning.